Validation patch for asset field types
Validate and hard-set casing on field_type values to protect against a known issue in the Hudu API.
In current versions of Hudu, the field_type value is case-sensitive and not validated server-side.  Sending a field_type value that is improperly cased can corrupt the asset, requiring manual removal from the database or restoring Hudu from a backup.

Relevent discussion in the MSPGeek Hudu Discord - https://discord.com/channels/801971115013963818/921032269546782720/1184556878689222716